backend/requirements.txt

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
werkzeug
pydantic-settings
pydantic[email]
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv

# File Processing
PyPDF2
python-docx
pandas
openpyxl

# Additional utilities
python-multipart
